Remove yaml converter
