+++ src/converter/__init__.py
@@ -1,2 +1,1 @@
-from .linkml_converter import *
-from .yaml_converter import *
+from .converter import *
